Refactor ABTests
The goal is to make the creation of AB tests (aka Experiment) easier and less error-prone.
Creating a Experiment would only require the following code in Experiments.scala.
An experiment is assigned to a existing participation group.
Pick one based on the size of the audience you want to assign to the experiment
```
object MyExperiment extends Experiment(
  name = "my-experiment",
  description = "A slice of the audience will participate in my
experiment"",
  owners = Seq(Owner.withGithub("tbonnin")),
  sellByDate = new LocalDate(2062, 2, 1),
  participationGroup = Perc1A
)

```

Requests would be bucketed in fixed groups in the VCL and frontend apps
would use the Vary response header to signify multiple variants exist
for the same request
Groups other than the 50% do NOT overlap to ensure experiments don't
influence each other.

This PR also creates automatically an opt-in/out endpoint for each
experiment